# docs/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: 出行安全卫士接口文档
schemes:
  - http
basePath: /api
paths:

  /login:
    post:
      summary: 登录
      operationId: login
      produces:
        - application/json
      tags:
        - login
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            type: object
            properties:
              account:
                type: string
                description: 登录账号
              password:
                type: string
                description: 密码
  /logout:
    post:
      summary: 退出
      operationId: logout
      produces:
        - application/json
      tags:
        - login
      responses:
        200:
          description: 成功
          schema:
            type: object
  /currentUser:
    get:
      summary: 获取当前用户
      operationId: currentUser
      produces:
        - application/json
      tags:
        - enter
      responses:
        200:
          description: 成功
          schema:
            type: object


  /user/get:
    get:
      summary: 获取单个用户信息
      operationId: getUserByQuery
      produces:
        - application/json
      tags:
        - user
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: userID
          in: query
          description: 用户ID
          type: integer
        - name: mobile
          in: query
          description: 用户手机号码
          type: string

  /user/list:
    get:
      summary: 获取用户列表信息
      operationId: listUser
      produces:
        - application/json
      tags:
        - user
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: type
          in: query
          description: 用户类型
          type: integer
        - name: gender
          in: query
          description: 用户性别
          type: integer
        - name: pageSize
          in: query
          description: 页数
          type: integer
        - name: page
          in: query
          description: 页码
          type: integer
  /user/register:
    post:
      summary: 用户注册接口
      operationId: registerUser
      produces:
        - application/json
      tags:
        - user
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            $ref: "#/definitions/UserCreateParam"

  /user/add/relationship:
    put:
      summary: 添加用户亲友
      operationId: addRelationship
      produces:
        - application/json
      tags:
        - user
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            $ref: "#/definitions/addRelationshipParam"


  /user/location/get:
    get:
      summary: 获取用户当前实时地址
      operationId: getLocation
      produces:
        - application/json
      tags:
        - userLocation
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: userID
          in: query
          description: 用户ID
          type: integer

  /user/location/list:
    get:
      summary: 获取用户当前实时地址
      operationId: listUserLocations
      produces:
        - application/json
      tags:
        - userLocation
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: userID
          in: query
          description: 用户ID
          type: integer

  /user/get/relationship:
    get:
      summary: 获取用户当前实时地址
      operationId: getRelationship
      produces:
        - application/json
      tags:
        - user
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: userID
          in: query
          description: 用户ID
          type: integer

  /user/location/set:
    put:
      summary: 更新用户实时地址
      operationId: setUserLocation
      produces:
        - application/json
      tags:
        - userLocation
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            $ref: "#/definitions/setUserLocationParam"


  /event/get:
    get:
      summary: 获取单个事件信息
      operationId: getEventByID
      produces:
        - application/json
      tags:
        - event
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: eventID
          in: query
          description: 事件ID
          type: integer

  /event/list:
    get:
      summary: 获取事件列表信息
      operationId: listEvent
      produces:
        - application/json
      tags:
        - event
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: token
          in: header
          description: token
          required: false
          type: string
        - name: eventID
          in: query
          description: 事件ID
          type: integer
        - name: userID
          in: query
          description: 报警人
          type: integer
        - name: fixerID
          in: query
          description: 处理人ID 警察
          type: integer
        - name: HelperID
          in: query
          description: 帮助人ID 保安 热心群众
          type: integer
        - name: status
          in: query
          description: 事件状态
          type: integer
        - name: level
          in: query
          description: 事件等级
          type: integer
        - name: pageSize
          in: query
          description: 页数
          type: integer
        - name: page
          in: query
          description: 页码
          type: integer

  /event/register:
    post:
      summary: 事件注册接口
      operationId: registerEvent
      produces:
        - application/json
      tags:
        - event
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            $ref: "#/definitions/EventCreateParam"


  /event/update:
    post:
      summary: 事件更新接口
      operationId: updateEvent
      produces:
        - application/json
      tags:
        - event
      responses:
        200:
          description: 成功
          schema:
            type: object
      parameters:
        - name: body
          in: body
          description: 请求参数
          schema:
            $ref: "#/definitions/EventUpdateParam"

# 引用参数
definitions:
  # 用户注册参数
  UserCreateParam:
    type: object
    properties:
      name:
        type: string
        description: 用户姓名
      mobile:
        type: string
        description: 手机号码
      password:
        type: string
        description: 用户登录密码 密码不能小于六位
      idCard:
        type: string
        description: 身份证号码 警员标志码
      gender:
        type: integer
        description: 性别 1-男 2-女
      age:
        type: integer
        description: 年龄
      type:
        type: integer
        description: 用户类型 1-用户 2-热心群众，其他安保人员 3-警察
      address:
        type: string
        description: 警察的话是警局 保安是负责街区 用户是常用活动范围
      longitude:
        type: number
        description: 经度
      latitude:
        type: number
        description: 维度
      photo:
        type: string
        description: 照片

  addRelationshipParam:
    type: object
    properties:
      relationship:
        type: string
        description: 用户亲友关系
      mobile:
        type: string
        description: 手机号码
      userID:
        type: integer
        description: 用户ID

  setUserLocationParam:
    type: object
    properties:
      userID:
        type: integer
        description: 用户ID
      longitude:
        type: number
        description: 身份证号码 警员标志码
      latitude:
        type: number
        description: 性别

  # 事件注册参数
  EventCreateParam:
    type: object
    properties:
      level:
        type: integer
        description: 手机号码
      address:
        type: string
        description: 事件登录密码
      longitude:
        type: number
        description: 身份证号码 警员标志码
      latitude:
        type: number
        description: 性别
      file:
        type: string
        description: 警察的话是警局 保安是负责街区 事件是常用活动范围
      userID:
        type: integer
        description: 用户ID

  EventUpdateParam:
    type: object
    properties:
      time:
        type: string
        description: 事件姓名
      level:
        type: integer
        description: 手机号码
      address:
        type: string
        description: 事件登录密码
      longitude:
        type: number
        description: 身份证号码 警员标志码
      latitude:
        type: number
        description: 性别
      status:
        type: integer
        description: 事件类型
      file:
        type: string
        description: 警察的话是警局 保安是负责街区 事件是常用活动范围
      userID:
        type: integer
        description: 用户ID
      fixerID:
        type: integer
        description: 警察ID
      helperID:
        type: integer
        description: 热心群众/保安ID